mtailor-test/requirements.txt
# Core Python libraries
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.1.0

# PyTorch and torchvision (for conversion and preprocessing)
torch==2.2.2
torchvision==0.17.2

# ONNX runtime for inference
onnxruntime==1.17.1
onnx==1.15.0

# Flask for creating the API server
flask==3.0.2

# Requests for testing the deployed API
requests==2.32.3

# Cerebrium for model management
cerebrium==1.40.4
